+++ CHANGELOG.md
@@ -5,10 +5,13 @@
 - [Feat][pull3963]: Allow customizing permission required to bypass max puppet limit (owllex)
 - [Fix][pull3922]: Fix MXP OPEN mode eating unknown tags like <username> in plain text (blongden)
 - [Fix][pull3962]: Prevent aborting channel message distribution when at_pre_channel_msg returns False/None (owllex)
+- [Fix][pull3968]: EvMenu: format dict helptext per entry (Sanjays2402)
+- Fix: EvMenu: render dict helptext as a readable list on 'help' instead of a raw dict (Griatch)
 
 [pull3922]: https://github.com/evennia/evennia/pull/3922
 [pull3962]: https://github.com/evennia/evennia/pull/3962
 [pull3963]: https://github.com/evennia/evennia/pull/3963
+[pull3968]: https://github.com/evennia/evennia/pull/3968
 
 ## Evennia 6.1.0